refactor: flatten common modules and update imports
