--- src/common/module/__init__.py
@@ -1,1 +1,0 @@
-"""Reusable common modules used across pipelines and services."""
