+++ src/core/storage_client.py
@@ -1,8 +1,8 @@
 """Compatibility shim for storage client imports.
 
-The implementation has been moved to `common.module.storage_client`.
+The implementation has been moved to `common.storage_client`.
 """
 
-from common.module.storage_client import RustFSClient, RustFSConfig
+from common.storage_client import RustFSClient, RustFSConfig
 
 __all__ = ["RustFSClient", "RustFSConfig"]
