+++ src/pipeline/scraper/module/base.py
@@ -1,8 +1,8 @@
 """Compatibility shim for scraper base classes.
 
-The implementation has been moved to `common.module.http_scraper`.
+The implementation has been moved to `common.http_scraper`.
 """
 
-from common.module.http_scraper import BaseHttpScraper, RequestSpec
+from common.http_scraper import BaseHttpScraper, RequestSpec
 
 __all__ = ["BaseHttpScraper", "RequestSpec"]
